requirements.txt
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dateutil==2.9.0
anthropic==0.49.0
